Request manager finished
