+++ src/main/java/frc/robot/Robot.java
@@ -24,7 +24,7 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import frc.robot.commands.RobotCommands;
 import frc.robot.stateMachine.OperatorOptions;
-import frc.robot.stateMachine.RobotManager;
+import frc.robot.stateMachine.RequestManager;
 import frc.robot.subsystems.climber.Climber;
 import frc.robot.subsystems.climber.WinchSpeeds;
 import frc.robot.subsystems.drivetrain.CommandSwerveDrivetrain;
@@ -33,7 +33,7 @@ public class Robot extends TimedRobot {
   private Command m_autonomousCommand;
 
   // Uncomment as needed
-  public static RobotManager robotManager = RobotManager.getInstance();
+  public static RequestManager robotManager = RequestManager.getInstance();
   public static RobotCommands robotCommands = new RobotCommands();
   public static final Controls controls = new Controls();
   private SendableChooser<Command> autoChooser;
