+++ src/main/java/frc/robot/stateMachine/RequestManagerStates.java
@@ -1,0 +1,15 @@
+package frc.robot.stateMachine;
+
+/*
+ * Naming Convention is: ACTION_LOCATION
+ */
+
+public enum RequestManagerStates {
+   INDEPENDENT,
+   PREPARE_HANDOFF, 
+   HANDOFF, 
+   PREPARE_INVERTED_HANDOFF, 
+   INVERTED_HANDOFF, 
+   CLIMB, 
+   PREPARE_IDLE 
+}
