--- src/main/java/frc/robot/stateMachine/RobotState.java
@@ -1,31 +1,0 @@
-package frc.robot.stateMachine;
-
-/*
- * Naming Convention is: ACTION_LOCATION
- */
-
-public enum RobotState {
-    INTAKING_CORAL,
-    PREPARE_HANDOFF,
-    HANDOFF,
-    IDLE,
-    WAIT_L1,
-    WAIT_L2,
-    WAIT_L3,
-    WAIT_L4,
-    SCORE_L1,
-    SCORE_L2,
-    SCORE_L3,
-    SCORE_L4,
-    GROUND_ALGAE_INTAKE,
-    HIGH_REEF_ALGAE_INTAKE,
-    LOW_REEF_ALGAE_INTAKE,
-    WAIT_BARGE,
-    WAIT_PROCESSOR,
-    SCORE_BARGE,
-    SCORE_PROCESSOR,
-    CLIMB,
-    RESET_TO_IDLE,
-    PREPARE_INVERTED_HANDOFF,
-    INVERTED_HANDOFF
-}
